Gef2Open veranderd in class aanroep
Functionaliteit is gelijk gebleven.
Library wordt nu gestart met:
import Gef2Open
myGef = Gef2Open.Gef2OpenClass()
myGef.read_gef('myfile.gef')
